Support akmods in rpm-ostree
When running in rpm-ostree systems (like silverblue) the %post
scripts run in a limited sandbox which produce a static image that
will be used as-is. This causes some issues for akmods.

First of all, we can't install the built kmod rpms or even modify /usr
at any time past the %post script.

Secondly, the sandbox puts some limits on what we can do in %post. For
example we can't use any users other than root (due to /usr being a fuse
mount), and /var is not available for either the rpm db or /var/cache.

So, as a result, when layering akmods on rpm-ostree we build the kmod
from the directly from the %post of the akmod rpm and extract it as
part of the final ostree image.

Note, this requires some changes to kmodtools to generate the call
to akmods-ostree-post in the akmod rpm.