akmods: Fix log timestamping logic
When copying the rpmbuild log (from akmodsbuild) to the kmod specific
log file, leading white space is not preserved (due to the read/echo
method used). As a result, the original indentation is lost, and perhaps
most importantly, the underlining of error locations by gcc is
misaligned. Additionally, '\' escapes are not preserved by the 'echo'
(e.g., "printf '%s\n'" comes out "printf '%sn'").

Instead of read/echo to append a timestamp and the program name, use
'sed'.

Fixes: fedora#2098266