Fix systemd-boot installed systems.
Fedora has a inst.sdboot option which installs kernels on the
ESP, it also provides a grubby shim for compatibility with
parts of the system wishing to have assistance in reading and
updating loader entries. This means that akmods tries to use
it and fails because it doesn't support the --default-kernel
switch. Adding that switch to the systemd/grubby shim results
in a path that looks like /boot/efi/XXXX/A.B.C-D00.fcXX.arch/linux
rather than the pattern akmods is using to detect the default kernel
entry. Meaning it needs to be tweaked to understand this anyway.

Instead, lets just fix akmods to understand the grubby symlink
and fix it to work with the existing bootctl logic. In this case
that involves telling jq to output "raw" unquoted strings so
that the rest of akmods is happy.

Signed-off-by: Jeremy Linton <jeremy.linton@arm.com>