Fix intel-ipu6-kmod installation with kernel >= 6.10
The intel-ipu6 kmod contains multiple .ko files. The main intel_ipu6.ko
as well as separate modules for the CSI-2 receiver (intel_ipu6_isys.ko)
and the ISP (intel_ipu6_psys.ko) functions.

The intel_ipu6.ko and intel_ipu6_isys.ko modules have been mainlined in
kernel 6.10. But the ISP (intel_ipu6_psys.ko) support has not been
mainlined and this will not be mainlined for a long time to come.

So the intel-ipu6 kmod is still useful for users to have.

But now that intel_ipu6.ko is part of the mainline kernel,
"modinfo -n intel-ipu6" will return a path to the version under
/lib/modules/$kver/kernel/ when the kmod is not installed yet
causing the check for no kmod being installed yet:

	# kmod present, even with weak-modules?
	if [[ ! -n "${kmodpackage_file}" ]] && [[ ! -d /lib/modules/${this_kernelver}/extra/${this_kmodname}/ ]] ; then

to fail, which in turn causes the version check to fail with:
"Warning: Could not determine what package owns /lib/modules/$kver/extra/intel-ipu6/"

Add a new special case for when "modinfo -n $kmodname" returns
a path under /lib/modules/$kver/kernel/ to fix this.